Fix segfault
This commit also switches to mpsc for communication across threads.
Right now it's limited in the way that equations will never be removed
or even modified, and duplicates are added each time you hit apply. A
more complicated way of fixing this would be have the mpsc channel send
an enum that can be Add or Remove, but we would need some kind of unique
identifier for equations. A simpler but less efficient way of doing it
would be having mpsc send a Vec of equations over every time.